Added templates
