+++ netbox_security/mixins/__init__.py
@@ -1,2 +1,1 @@
 from .FirewallFilterRule import *
-from .FirewallRuleSetting import *
